exposing waveform function for AudioBuffer
