+++ lib/builders/webaudio.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var getAudioDecoder = require("./audiodecoder");
+var getOptions = require('./options');
 
 /**
  * Creates a working WaveformData based on binary audio data.
@@ -42,31 +43,13 @@ var getAudioDecoder = require("./audiodecoder");
 
 function fromAudioObjectBuilder(audio_context, raw_response, options, callback) {
   var audioContext = window.AudioContext || window.webkitAudioContext;
-  var defaultOptions = {
-    scale: 512,
-    amplitude_scale: 1.0
-  };
-
   if (!(audio_context instanceof audioContext)) {
     throw new TypeError("First argument should be an AudioContext instance");
   }
 
-  // fromAudioObjectBuilder(audioContext, data, callback) form
-  if (typeof options === "function") {
-    callback = options;
-    options = {};
-  }
-  else {
-    options = options || {};
-  }
-
-  options.scale = options.scale || defaultOptions.scale;
-  options.amplitude_scale = options.amplitude_scale || defaultOptions.amplitude_scale;
-
-  if (options.hasOwnProperty("scale_adjuster")) {
-    throw new Error("Please rename the 'scale_adjuster' option to 'amplitude_scale'");
-  }
-
+  var opts = getOptions(options, callback);
+  callback = opts.callback;
+  options = opts.options;
   // The following function is a workaround for a Webkit bug where decodeAudioData
   // invokes the errorCallback with null instead of a DOMException.
   // See https://webaudio.github.io/web-audio-api/#dom-baseaudiocontext-decodeaudiodata
